moved all e2e tests to ./tests directory
